scheduler: refactor: split build.rs into lifecycle, self-heal, waiting-state, and buildability modules
